nextjs-flask/api/requirements.txt
cachetools==5.4.0
Flask==3.0.3
lancedb==0.15.0
langchain_community==0.3.7
langchain_openai==0.2.8
python-dotenv==1.0.1
